workload: whitelist allowable dependencies
We want pkg/workload to be lightweight and go gettable, mostly common
interfaces and data structures.

This commit adds a test to lint the transitive dependency graph of
pkg/workload against a whitelist. It also moves out some helpers to a
new package to keep this whitelist sane. Everything in the new
workloadsql package is purely code movement, no functional change.

In this case, the helpers take in workload structs and use them on a
CockroachDB instance via SQL. By a similar argument, we should move the
CSV stuff out and perhaps even split up workload's ideas of dataset and
queryload generators, but for now we make a compromise between
invasiveness and dep health.

Release note: None
