+++ pkg/workload/bank/bank_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
-	"github.com/cockroachdb/cockroach/pkg/workload"
+	"github.com/cockroachdb/cockroach/pkg/workload/workloadsql"
 )
 
 func TestBank(t *testing.T) {
@@ -53,7 +53,7 @@ func TestBank(t *testing.T) {
 			bankTable := bank.Tables()[0]
 			sqlDB.Exec(t, fmt.Sprintf(`CREATE TABLE %s %s`, bankTable.Name, bankTable.Schema))
 
-			if err := workload.Split(ctx, db, bankTable, 1 /* concurrency */); err != nil {
+			if err := workloadsql.Split(ctx, db, bankTable, 1 /* concurrency */); err != nil {
 				t.Fatalf("%+v", err)
 			}
