+++ pkg/workload/workload_test.go
@@ -13,19 +13,11 @@
 package workload_test
 
 import (
-	"context"
-	"fmt"
-	"strconv"
 	"testing"
 
-	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
-	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
-	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
-	"github.com/cockroachdb/cockroach/pkg/util/uuid"
 	"github.com/cockroachdb/cockroach/pkg/workload"
-	"github.com/cockroachdb/cockroach/pkg/workload/bank"
 )
 
 func TestGet(t *testing.T) {
@@ -35,127 +27,6 @@ func TestGet(t *testing.T) {
 	}
 	if _, err := workload.Get(`nope`); !testutils.IsError(err, `unknown generator`) {
 		t.Errorf(`expected "unknown generator" error got: %+v`, err)
-	}
-}
-
-func TestSetup(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	tests := []struct {
-		rows        int
-		batchSize   int
-		concurrency int
-	}{
-		{10, 1, 1},
-		{10, 9, 1},
-		{10, 10, 1},
-		{10, 100, 1},
-		{10, 1, 4},
-		{10, 9, 4},
-		{10, 10, 4},
-		{10, 100, 4},
-	}
-
-	ctx := context.Background()
-	s, db, _ := serverutils.StartServer(t, base.TestServerArgs{UseDatabase: `test`})
-	defer s.Stopper().Stop(ctx)
-	sqlutils.MakeSQLRunner(db).Exec(t, `CREATE DATABASE test`)
-
-	for _, test := range tests {
-		t.Run(fmt.Sprintf("rows=%d/batch=%d", test.rows, test.batchSize), func(t *testing.T) {
-			sqlDB := sqlutils.MakeSQLRunner(db)
-			sqlDB.Exec(t, `DROP TABLE IF EXISTS bank`)
-
-			gen := bank.FromRows(test.rows)
-			if _, err := workload.Setup(ctx, db, gen, test.batchSize, test.concurrency); err != nil {
-				t.Fatalf("%+v", err)
-			}
-
-			for _, table := range gen.Tables() {
-				var c int
-				sqlDB.QueryRow(t, fmt.Sprintf(`SELECT count(*) FROM %s`, table.Name)).Scan(&c)
-				if c != table.InitialRows.NumTotal {
-					t.Errorf(`%s: got %d rows expected %d`,
-						table.Name, c, table.InitialRows.NumTotal)
-				}
-			}
-		})
-	}
-}
-
-func TestSplits(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	ctx := context.Background()
-	s, db, _ := serverutils.StartServer(t, base.TestServerArgs{UseDatabase: `test`})
-	defer s.Stopper().Stop(ctx)
-	sqlutils.MakeSQLRunner(db).Exec(t, `CREATE DATABASE test`)
-
-	for _, ranges := range []int{1, 2, 3, 4, 10} {
-
-		tables := []workload.Table{
-			{
-				Name:   `ints`,
-				Schema: `(a INT PRIMARY KEY)`,
-				Splits: workload.Tuples(ranges-1, func(i int) []interface{} {
-					return []interface{}{i}
-				}),
-			},
-			{
-				Name:   `floats`,
-				Schema: `(a FLOAT PRIMARY KEY)`,
-				Splits: workload.Tuples(ranges-1, func(i int) []interface{} {
-					return []interface{}{float64(i)}
-				}),
-			},
-			{
-				Name:   `strings`,
-				Schema: `(a STRING PRIMARY KEY)`,
-				Splits: workload.Tuples(ranges-1, func(i int) []interface{} {
-					return []interface{}{strconv.Itoa(i)}
-				}),
-			},
-			{
-				Name:   `bytes`,
-				Schema: `(a BYTES PRIMARY KEY)`,
-				Splits: workload.Tuples(ranges-1, func(i int) []interface{} {
-					return []interface{}{strconv.Itoa(i)}
-				}),
-			},
-			{
-				Name:   `uuids`,
-				Schema: `(a UUID PRIMARY KEY)`,
-				Splits: workload.Tuples(ranges-1, func(i int) []interface{} {
-					u, err := uuid.NewV4()
-					if err != nil {
-						panic(err)
-					}
-					return []interface{}{u.String()}
-				}),
-			},
-		}
-
-		t.Run(fmt.Sprintf("ranges=%d", ranges), func(t *testing.T) {
-			sqlDB := sqlutils.MakeSQLRunner(db)
-			for _, table := range tables {
-				sqlDB.Exec(t, fmt.Sprintf(`DROP TABLE IF EXISTS %s`, table.Name))
-				sqlDB.Exec(t, fmt.Sprintf(`CREATE TABLE %s %s`, table.Name, table.Schema))
-
-				const concurrency = 10
-				if err := workload.Split(ctx, db, table, concurrency); err != nil {
-					t.Fatalf("%+v", err)
-				}
-
-				countRangesQ := fmt.Sprintf(
-					`SELECT count(*) FROM [SHOW EXPERIMENTAL_RANGES FROM TABLE test.%s]`, table.Name,
-				)
-				var actual int
-				sqlDB.QueryRow(t, countRangesQ).Scan(&actual)
-				if ranges != actual {
-					t.Errorf(`expected %d got %d`, ranges, actual)
-				}
-			}
-		})
 	}
 }
